# go/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Bes Ops API
  description: Uptime monitoring API
  version: "1.0.0"
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      x-ogen-operation-group: Auth
      security: []
      summary: Authenticate with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /auth/token/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      x-ogen-operation-group: Auth
      summary: Refresh JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: New token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      description: Standard error response returned for all 4xx and 5xx responses.
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          description: Machine-readable error code following gRPC status conventions (e.g. NOT_FOUND, INVALID_ARGUMENT, INTERNAL).
          enum:
            - OK
            - CANCELED
            - UNKNOWN
            - INVALID_ARGUMENT
            - DEADLINE_EXCEEDED
            - NOT_FOUND
            - ALREADY_EXISTS
            - PERMISSION_DENIED
            - RESOURCE_EXHAUSTED
            - FAILED_PRECONDITION
            - ABORTED
            - OUT_OF_RANGE
            - UNIMPLEMENTED
            - INTERNAL
            - UNAVAILABLE
            - DATA_LOSS
            - UNAUTHENTICATED
        error:
          type: string
          maxLength: 500
          description: Human-readable error message describing what went wrong.
...
    LoginRequest:
      description: Credentials for authenticating a user session.
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64
          description: The account username.
        password:
          type: string
          format: password
          minLength: 6
          maxLength: 255
          description: The account password.
        token:
          type: string
          minLength: 6
          maxLength: 6
          pattern: "^[0-9]{6}$"
          description: 2FA TOTP token. Required if the account has two-factor authentication enabled.
    LoginResponse:
      description: Response after a successful login attempt.
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: JWT bearer token for authenticating subsequent requests. Signed with ECDSA (ES256).
        refreshToken:
          type: string
          description: Opaque refresh token for obtaining new access tokens without re-authenticating.
        tokenRequired:
          type: boolean
          description: If true, the account has 2FA enabled and a TOTP token must be provided to complete login.
    RefreshTokenRequest:
      description: Request body for refreshing an access token.
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          description: The refresh token issued during login.
    TokenResponse:
      description: Response containing a fresh JWT token.
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Newly issued JWT bearer token.
...